+++ python/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the Python SDK package `marpledata` will be documented in
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0] - 2026-08-30
+
+### Added
+
+- `DB.delete_signals`, `Dataset.delete_signal` / `Dataset.delete_signals`, and `Signal.delete` to remove signals from a dataset.
+- Allow `add_signal` in post-processing
+
 ## [3.5.0] - 2026-08-20
 
 ### Added